Title: 자연수 n개의 합, Time: 75ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 231211/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -15,7 +15,7 @@ public class Main {
     }
     
     static void init()throws IOException{
-    	limit = Integer.valueOf(buffer.readLine());
+    	limit = Long.valueOf(buffer.readLine());
     }
